N beads are resting on a smooth horizontal wire...
n beads are resting on a smooth horizontal wire which is circular at the end with radius R as shown. The masses of the beads are m,m/2,m/4,... m/2^(n-1).respectively. Find the minimum velocity that should be imparted to the first bead of mass m such that the nth bead will fall in the tank shown in the figure. Assume all collisions to be elastic.