Find velocity of the Wedge(M) when the block(m) reaches the bottom..
A block of mass m is placed on a triangular block of mass M, which in turn is placed on  horizontal surface as shown in figure (9-E21). Assuming frictionless surfaces find the velocity of the triangular block when the smaller block reaches the bottom end.
or
A wedge of mass M rests on an absolutely smooth horizontal surface. A block of mass m is placed on wedge inclined at an angle alpha find velocity of M when the block falls through a height h.