A small ring of mass m can slide on a smooth circular wire of radius r and center O, which is fixed in the vertical plane. From a point on the wire at a vertical distance 1/2 above, the ring is given a velocity sqrt(gr) along the downward tangent to the wire. Show that it will just reach the highest point of the wire. Find the reaction between ring and wire when the ring is at a vertical distance 1/2 below, O.
A small ring of mass m can slide on a smooth circular wire of radius r and center O, which is fixed in the vertical plane. From a point on the wire at a vertical distance 1/2 above, the ring is given a velocity sqrt(gr) along the downward tangent to the wire. Show that it will just reach the highest point of the wire. Find the reaction between ring and wire when the ring is at a vertical distance 1/2 below, O.